Crowd control system
CC state is stored in an int, using bitmasks to query whether or not a
unit's specific capabilities are impaired. For more info see my post in
dev_docs on the discord.

CC effects are stored in an array, and will only remove a CC type on
effect end if there is not already another CC effect applying it.